abq step classes out of exmpl file
